+++ src/components/lottery/lottery-result/lottery-result-header/LotteryResultHeader.vue
@@ -1,13 +1,14 @@
 <template>
-  <div class="row">
-    <div class="col-md-12">
-      <h2 class="text-center">{{ resultTitle }}</h2>
+  <div class="container">
+    <div class="row">
+      <div class="col-md-12">
+        <h2 class="text-center">{{ title }}</h2>
+      </div>
     </div>
   </div>
 </template>
 
 <script>
-import {onMounted, ref, watch} from "vue";
 
 export default {
   props: {
@@ -16,19 +17,6 @@ export default {
       required: false,
       default: 'Wyniki z poprzedniej loterii'
     }
-  },
-  setup(props) {
-    const resultTitle = ref('');
-
-    onMounted(() => {
-      resultTitle.value = props.title;
-    })
-
-    watch(() => props.title, () => {
-      resultTitle.value = props.title;
-    });
-
-    return {resultTitle}
   }
 }
 </script>
